Refactored directories. Added CmakeFile.
