+++ src/HistogramSampler.h
