+++ src/LogHistogramSampler.h
