+++ src/barrier.cc
