+++ src/binary_protocol.h
