+++ src/cmdline.ggo
@@ -65,6 +65,10 @@ option "scan" - "Scan latency across QPS rates from min to max."
 text "\nAgent-mode options:"
 option "agentmode" A "Run client in agent mode."
 option "masterless_agentmode" M "Run client in masterless agent mode."
+option "sin" - "[Only in masterless mode] Run sinusoidal workload with \
+specified ms period." int default="500"
+option "push_lat_influx" - "[Only in masterless mode] Push latencies to \
+influxDB on specified address." string typestr="host:port"
 option "agent" a "Enlist remote agent." string typestr="host" multiple
 option "agent_port" p "Agent port." string default="5556"
 option "lambda_mul" l "Lambda multiplier.  Increases share of \
